pyproject.toml
[project]
name = "rover-client"
version = "0.1.0"
description = "Firmware for my Lege bot"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncio>=4.0.0",
    "pimoroni-bme280>=1.0.0",
    "python-periphery>=2.4.1",
    "ruff>=0.14.8",
    "smbus>=1.1.post2",
]
